version 1.02 - 거래 명세서 정보 수정 - 숫자 정보 콤마 추가 - 외부 저장소에 저장하기 위한 인텐트 설정 - 한글 사용을 위한 폰트 설정 및 적용 (배민 한나프로)

+++ app/src/main/java/com/skymilk/androidpdfwriter/Invoice.kt
@@ -18,6 +18,6 @@ data class PersonInfo(
 
 data class Product(
     val description: String,
-    val rate: Float,
+    val rate: Int,
     val quantity: Int
 )
